refactor: the spec, the null peripherals and the builder share one file
Two files of mostly imports were one small module's worth of code.
